pyproject.toml
[project]
name = "nome_do_projeto"
version = "0.1.0"
description = "Descrição do projeto"
authors = [
    { name="Seu Nome", email="seu.email@example.com" }
]
# dependencies = [
#    django = "*",
#    gunicorn = "*",
#    uvicorn = {extras = ["standard"], version = "*"},
#    psycopg2-binary = "*",
#    python-dotenv = "*",
#]
requires-python = "==3.12.*"
dependencies = [
    "django>=5.1.3",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "dj-database-url>=2.3.0",
    "pillow>=11.0.0",
    "python-magic>=0.4.27; sys_platform == \"linux\"",
    "python-magic-bin>=0.4.14; sys_platform == \"win32, darwin\"",
    "drf-yasg>=1.21.8",
    "django-restframework>=0.0.1",
    "uvicorn>=0.32.1",
    "poetry>=1.8.4",
    "djangorestframework-simplejwt>=5.3.1",
    "whitenoise>=6.8.2",
]


[tool.pdm]
distribution = false

[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"


[tool.pdm.scripts]
pre_dev = "python ./scripts/set_my_ip.py"
dev = "python manage.py runserver 0.0.0.0:19003"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"
post_migrate = "python manage.py graph_models --disable-sort-fields -g -o core.png core uploader"



[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.backend"

